+++ infrastructure/src/test/java/com/frb/infrastructure/models/PurchaseRequestTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.json.JacksonTester;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 
@@ -20,7 +21,7 @@ public class PurchaseRequestTest {
     public void testMarshall() throws Exception {
         final var expectedDescription = "Purchase Description";
         final var purchaseDate = LocalDate.now();
-        final var amount = Double.valueOf(11.47);
+        final var amount = BigDecimal.valueOf(11.47);
 
         final var request =
                 new PurchaseRequest(expectedDescription, purchaseDate, amount);
@@ -37,7 +38,7 @@ public class PurchaseRequestTest {
     public void testUnmarshall() throws Exception {
         final var expectedDescription = "Purchase Description";
         final var purchaseDate = LocalDate.now();
-        final var amount = Double.valueOf(11.47);
+        final var amount = BigDecimal.valueOf(11.47);
 
         final var json = """
                 {
